refactor(mirage): split 1361-line createMirage.ts into focused modules
Extract types, MirageCore class, hydration logic, and proxy factories
into separate files with clear single responsibilities:

- mirage.types.ts: All type definitions and exported constants
- MirageCore.ts: State machine class and plugin helpers
- hydration.ts: Event replay/hydration logic
- proxy/collectionProxy.ts: List/entity proxy factories
- proxy/mapProxy.ts: Map proxy factories
- proxy/selectorProxy.ts: Selector invocation and result wrapping
- proxy/deepProxy.ts: Root proxy factory and context wiring
- proxy/proxyContext.ts: Shared context interface

No file exceeds 400 lines. Public API unchanged.
All 62 existing tests pass without modification.
